# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sevasrot - Cow Welfare Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterBody {
  name: string;
  email: string;
  password: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  user: "user",
  admin: "admin",
} as const;

export type UserAuthProvider =
  (typeof UserAuthProvider)[keyof typeof UserAuthProvider];

export const UserAuthProvider = {
  local: "local",
  google: "google",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  authProvider: UserAuthProvider;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface CreateDonationBody {
  amount: number;
  isAnonymous: boolean;
  /** @nullable */
  displayName?: string | null;
}

export type DonationStatus =
  (typeof DonationStatus)[keyof typeof DonationStatus];

export const DonationStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface Donation {
  id: number;
  userId: number;
  amount: number;
  isAnonymous: boolean;
  /** @nullable */
  displayName?: string | null;
  status: DonationStatus;
  createdAt: string;
  /** @nullable */
  reviewedAt?: string | null;
  /** @nullable */
  reviewedBy?: number | null;
}

export type DonationWithUserStatus =
  (typeof DonationWithUserStatus)[keyof typeof DonationWithUserStatus];

export const DonationWithUserStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface DonationWithUser {
  id: number;
  userId: number;
  amount: number;
  isAnonymous: boolean;
  /** @nullable */
  displayName?: string | null;
  status: DonationWithUserStatus;
  createdAt: string;
  /** @nullable */
  reviewedAt?: string | null;
  /** @nullable */
  reviewedBy?: number | null;
  /** @nullable */
  userName?: string | null;
  /** @nullable */
  userEmail?: string | null;
}

export interface ApprovedDonationsResponse {
  donations: Donation[];
  total: number;
}

export interface SevaDrive {
  id: number;
  title: string;
  description: string;
  location: string;
  date: string;
  images: string[];
  createdBy: number;
  createdAt: string;
}

export interface CreateDriveBody {
  title: string;
  description: string;
  location: string;
  date: string;
}

export interface UpdateDriveBody {
  title?: string;
  description?: string;
  location?: string;
  date?: string;
  images?: string[];
}

export interface PendingCountResponse {
  count: number;
}