# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sevasrot - Cow Welfare Platform API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: donations
    description: Donation operations
  - name: drives
    description: Seva Drive operations
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user info
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /donations/create:
    post:
      operationId: createDonation
      tags: [donations]
      summary: Create a new donation (pending)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDonationBody"
      responses:
        "201":
          description: Donation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donation"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /donations/approved:
    get:
      operationId: getApprovedDonations
      tags: [donations]
      summary: Get all approved donations with total
      responses:
        "200":
          description: Approved donations list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovedDonationsResponse"

  /donations/my:
    get:
      operationId: getMyDonations
      tags: [donations]
      summary: Get current user's donations
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User's donations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Donation"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /donations/pending:
    get:
      operationId: getPendingDonations
      tags: [donations]
      summary: Get all pending donations (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Pending donations list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DonationWithUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /donations/{id}/approve:
    patch:
      operationId: approveDonation
      tags: [donations]
      summary: Approve a donation (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Donation approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donation"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /donations/{id}/reject:
    patch:
      operationId: rejectDonation
      tags: [donations]
      summary: Reject a donation (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Donation rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donation"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /drives:
    get:
      operationId: getAllDrives
      tags: [drives]
      summary: Get all seva drives
      responses:
        "200":
          description: List of seva drives
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SevaDrive"

  /drives/{id}:
    get:
      operationId: getDrive
      tags: [drives]
      summary: Get a single seva drive
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Seva drive details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SevaDrive"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateDrive
      tags: [drives]
      summary: Update a seva drive (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDriveBody"
      responses:
        "200":
          description: Updated seva drive
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SevaDrive"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteDrive
      tags: [drives]
      summary: Delete a seva drive (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /drives/create:
    post:
      operationId: createDrive
      tags: [drives]
      summary: Create a new seva drive with image upload (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CreateDriveBody"
      responses:
        "201":
          description: Seva drive created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SevaDrive"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/pending-count:
    get:
      operationId: getPendingCount
      tags: [admin]
      summary: Get count of pending donations
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Pending donations count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PendingCountResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/all-donations:
    get:
      operationId: getAllDonations
      tags: [admin]
      summary: Get all donations (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All donations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DonationWithUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
      required:
        - name
        - email
        - password

    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password

    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user

    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: ["user", "admin"]
        authProvider:
          type: string
          enum: ["local", "google"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - authProvider
        - createdAt

    CreateDonationBody:
      type: object
      properties:
        amount:
          type: number
        isAnonymous:
          type: boolean
        displayName:
          type: ["string", "null"]
      required:
        - amount
        - isAnonymous

    Donation:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        amount:
          type: number
        isAnonymous:
          type: boolean
        displayName:
          type: ["string", "null"]
        status:
          type: string
          enum: ["pending", "approved", "rejected"]
        createdAt:
          type: string
          format: date-time
        reviewedAt:
          type: ["string", "null"]
          format: date-time
        reviewedBy:
          type: ["integer", "null"]
      required:
        - id
        - userId
        - amount
        - isAnonymous
        - status
        - createdAt

    DonationWithUser:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        amount:
          type: number
        isAnonymous:
          type: boolean
        displayName:
          type: ["string", "null"]
        status:
          type: string
          enum: ["pending", "approved", "rejected"]
        createdAt:
          type: string
          format: date-time
        reviewedAt:
          type: ["string", "null"]
          format: date-time
        reviewedBy:
          type: ["integer", "null"]
        userName:
          type: ["string", "null"]
        userEmail:
          type: ["string", "null"]
      required:
        - id
        - userId
        - amount
        - isAnonymous
        - status
        - createdAt

    ApprovedDonationsResponse:
      type: object
      properties:
        donations:
          type: array
          items:
            $ref: "#/components/schemas/Donation"
        total:
          type: number
      required:
        - donations
        - total

    SevaDrive:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        location:
          type: string
        date:
          type: string
          format: date-time
        images:
          type: array
          items:
            type: string
        createdBy:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - location
        - date
        - images
        - createdBy
        - createdAt

    CreateDriveBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        location:
          type: string
        date:
          type: string
      required:
        - title
        - description
        - location
        - date

    UpdateDriveBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        location:
          type: string
        date:
          type: string
        images:
          type: array
          items:
            type: string

    PendingCountResponse:
      type: object
      properties:
        count:
          type: integer
      required:
        - count