# lib/api-zod/src/generated/types/createDonationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sevasrot - Cow Welfare Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateDonationBody {
  amount: number;
  isAnonymous: boolean;
  /** @nullable */
  displayName?: string | null;
}