# lib/api-zod/src/generated/types/createDriveBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sevasrot - Cow Welfare Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateDriveBody {
  title: string;
  description: string;
  location: string;
  date: string;
}