# lib/api-zod/src/generated/types/donation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sevasrot - Cow Welfare Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { DonationStatus } from "./donationStatus";

export interface Donation {
  id: number;
  userId: number;
  amount: number;
  isAnonymous: boolean;
  /** @nullable */
  displayName?: string | null;
  status: DonationStatus;
  createdAt: Date;
  /** @nullable */
  reviewedAt?: Date | null;
  /** @nullable */
  reviewedBy?: number | null;
}