# lib/api-zod/src/generated/types/donationWithUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sevasrot - Cow Welfare Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { DonationWithUserStatus } from "./donationWithUserStatus";

export interface DonationWithUser {
  id: number;
  userId: number;
  amount: number;
  isAnonymous: boolean;
  /** @nullable */
  displayName?: string | null;
  status: DonationWithUserStatus;
  createdAt: Date;
  /** @nullable */
  reviewedAt?: Date | null;
  /** @nullable */
  reviewedBy?: number | null;
  /** @nullable */
  userName?: string | null;
  /** @nullable */
  userEmail?: string | null;
}