# lib/api-zod/src/generated/types/updateDriveBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sevasrot - Cow Welfare Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateDriveBody {
  title?: string;
  description?: string;
  location?: string;
  date?: string;
  images?: string[];
}